feat: ejs, add new error handling, code splitting
