+++ src/zip/files/fileToCompress.txt
@@ -1,0 +1,1 @@
+Compress me!
